Adding the GenKBot.cs - following the lead from the Icebreaker App Template code structure
